+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 ## 2.0.5
 
+- Minor: Server perks are now sorted alphabetically. (#711)
 - Fix: Release script not working. (#704)
+- Dev: Added unit tests for the Discord invite resolver. (#711)
 
 ## 2.0.4