+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Minor: Server perks are now sorted alphabetically. (#711)
+- Minor: Add stream information to twitch users. (#714)
 - Dev: Added unit tests for the Discord invite resolver. (#711)
 - Dev: Replace github.com/golang/mock with go.uber.org/mock. (#712)
 - Dev: Ignore mock files when generating code coverage. (#713)