+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Minor: Add stream information to twitch users. (#714)
 - Minor: Server perks are now sorted alphabetically. (#711)
 - Minor: Use correct domains for twitch user resolver. (#715)
+- Fix: Some Twitch clips being parsed as Twitch user profiles. (#755)
 - Dev: Changed the default `twitch-username-cache-duration` from 1 hour to 10 minutes. (#714)
 - Dev: Added unit tests for the Discord invite resolver. (#711)
 - Dev: Replace github.com/golang/mock with go.uber.org/mock. (#712)